Swap std::string with const char* in most places
Refactor Stats to expect const char* instead of std::string so we don't have to cast and pass string literals ("Yes"/"No") as std::strings.

In a proper implementation, the possible classes could be in an array of std::strings, but for this simple example with only two classes, literals are enough.